sim new api on net
downstream refs:
de6e3a99 afa66b0c 306313ba 8ace53af 3b7572a2 b02f2dfb ad1d4b5c 40df54d5 cc4b9921 187eadc4 1fc37d40 b50c0ffd 97886001 d5830468 2960666f 44f31611 ade183bf 3df22f92 7f839f75 5fc47d2f 739ea9a0 38f842b1 fe5abac2 059a0fa4 542ad444 804c710d 362c54b5 98f0357b f699d6e7 35405b10 5548e29a 5de64943 f7572db3 6e051e3d

+++ core0/us/gov/id/types.h
@@ -27,7 +27,7 @@ namespace us {
 
     using pport_t = port_t;
     using pin_t = uint16_t;
-    using version_fingerprint_t = uint16_t;
+    using version_fingerprint_t = uint8_t;
 
 }
